// sanjeevani-console/src/constants/global/enums.ts
export enum ROLES {
  ADMIN = "ADMIN",
  DOCTOR = "DOCTOR",
  NURSE = "NURSE",
}

export enum PERMISSION_MODULES {
  ADMINISTRATION = "ADMINISTRATION",
  ANALYTICS = "ANALYTICS",
  PATIENT_MANAGEMENT = "PATIENT_MANAGEMENT",
}

export enum PERMISSION_SUB_MODULES {
  AUDIT_LOGS = "AUDIT_LOGS",
  DASHBOARD = "DASHBOARD",
  ENCOUNTERS_MANAGEMENT = "ENCOUNTERS_MANAGEMENT",
  PATIENTS_MANAGEMENT = "PATIENTS_MANAGEMENT",
  USERS_MANAGEMENT = "USERS_MANAGEMENT",
}

export enum PERMISSIONS {
  READ_ALL = "READ_ALL",
  READ_OWNED = "READ_OWNED",
  WRITE_ALL = "WRITE_ALL",
  WRITE_OWNED = "WRITE_OWNED",
}

export enum GENDER_TYPES {
  FEMALE = "FEMALE",
  MALE = "MALE",
  OTHER = "OTHER",
}

export enum ENCOUNTER_SEVERITY {
  CRITICAL = "CRITICAL",
  MILD = "MILD",
  MODERATE = "MODERATE",
  SEVERE = "SEVERE",
}

export enum ENCOUNTER_STATUS {
  CLOSED = "CLOSED",
  FOLLOW_UP_REQUIRED = "FOLLOW_UP_REQUIRED",
  REFERRED = "REFERRED",
  UNDER_TREATMENT = "UNDER_TREATMENT",
}

export enum ENCOUNTER_VISIT_TYPES {
  EMERGENCY = "EMERGENCY",
  FOLLOW_UP = "FOLLOW_UP",
  NEW_CONSULTATION = "NEW_CONSULTATION",
  TELECONSULTATION = "TELECONSULTATION",
}

export enum EXPORT_FORMATS {
  CSV = "CSV",
  JSON = "JSON",
  XLS = "XLS",
}

export enum USER_STATUS_TYPES {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
}